Changes on IndexFactory and main.cpp
IndexFactory now uses algorithmIDs instead of algorithmNames to
instantiate the algorithms. Changed the if/else for switch/case, this
has the nice benefit that if two algorithms have the same ID because of
a programmer mistake then that switch case will detect it.

Main doesn't run one algorithm anymore, it runs all of them.
Main also saves all the results in a single csv file.