Changing how to keep track of measurements
Instead of having hard-coded vectors that each index needs to append to
in eavery measurement, now the Measurements is a map from a string to a
vector of strings.
This makes it easier to quickly add new measurement to each index
without having to change the code to write it to file.