Fixing split string and a bunch of warnings
- Fixed SplitString helper method to actually use the delimiter
- Fixed SplitString to actually convert to int64_t or float or string.
These are the available conversions, if more are necessary then
simply implement them in the template method from_string in the
class.
- Fixed multiple warnings about comparisons between size_t and int64_t.
- Changed multiple variables that were int64_t but should have been
size_t. Examples are: in the KDNode the position, column. In the search
algorhithms instead of return a vector of pairs of int64_t now it
returns a vector of pairs of size_t. And more...